refactor(core): extract update download and apply pipeline helpers
